+++ dashboard/app.py
@@ -3,8 +3,8 @@
 import os
 import threading
 from dotenv import load_dotenv
-from core.server import start_server
-from core.iothub import receive_events_from_iothub
+from .core.server import start_server
+from .core.iothub import receive_events_from_iothub
 
 load_dotenv()
